Moved bunkr URL fixing to scraper
